Results window improvements
I worked quite a lot with this, without success. My idea was to
overlay a "property toolbox" over the graph when the cursor was on
a bar. I tried to make a new Container in which a cairo context
could be used to draw something, but I was unable to put child
widgets on top. I also tried to use GooCanvas library, but when
everything worked well with a simple button, I noticed that it wont
be the same for something more complicated (in fact I've been warned
about limitation of Goocanvas::Widget, but I didn't understand
if it could fit my needs). I've spent too much time in this problem,
now I simply added a HPaned and put some Pocket information in the
second child. I don't like that, but for now I think it's the best
solution. When Ubuntu and derived distributions will be using
Gtk >= 3.2 I could -- maybe -- implement Overlay to obtain what I
wanted.