Changed MetaStream behaviour
Now MetaStream is a template class. A complete implementation is needed to let it work with basic_iostream, but it will be done soon. Oh, naturally everything inside MetaStream.cpp have been removed and put inside the header file, because of the problem related to template classes and file splitting. For now I simply changed implementations from MetaStream to MetaStream<ifstream>, but I want to change Session behaviour, so it can be used also to write session files.