Session is now a template class
This modification allow to handle MetaStream and sessions in a more dynamic way. Now if we need a read-only session, we can use a Session<ifstream>, when a read-write session is needed a Session<fstream> can be declared. Notice that this "little" change brings to a different handling of other classes like Gromacs and SasAnalysis. In fact SasAnalysis is now a template class too, and Gromacs handles ofstreams (with Session<ofstream>) when calculateSas is called. Everything could be done thanks to MetaStream.