Fixed problem with concatenated trajectories
A very huge bug still exists, but for now I have no idea how to solve it.
When trajectories are concatenated it is possible to change the frame "time parameter" to transform two dynamics 1-10;1-6 in one dynamic 1-16. This commit allows to analyse these kind of concatenated dynamics if the timestep is constant. If not so, or if the time parameter is left unchanged I have no idea how to distinguish this kind of trajectory from a "normal" one. And the program hangs.